Move Cell components to dedicated directory
