+++ gui/src/renderer/components/RelayStatusIndicator.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import styled from 'styled-components';
 import { colors } from '../../config.json';
-import * as Cell from './Cell';
+import * as Cell from './cell';
 
 const StyledRelayStatus = styled.div((props: { active: boolean }) => ({
   width: '16px',
