+++ gui/src/renderer/components/cell/Section.tsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import styled from 'styled-components';
+import { colors } from '../../../config.json';
+import { buttonText } from '../common-styles';
+
+const StyledSection = styled.div({
+  display: 'flex',
+  flexDirection: 'column',
+});
+
+export const SectionTitle = styled.span(buttonText, {
+  backgroundColor: colors.blue,
+  padding: '14px 16px 14px 22px',
+  marginBottom: '1px',
+});
+
+export const CellSectionContext = React.createContext<boolean>(false);
+
+export function Section(props: React.HTMLAttributes<HTMLDivElement>) {
+  const { children, ...otherProps } = props;
+  return (
+    <StyledSection {...otherProps}>
+      <CellSectionContext.Provider value={true}>{children}</CellSectionContext.Provider>
+    </StyledSection>
+  );
+}
